+++ Angular/Udemy/SUBHASISH/MEAN Stack/MEAN-STACK-LEARNING/Server/Controller/posts.js
@@ -1,0 +1,113 @@
+const Post = require("../models/post");
+
+exports.createPosts = (req, res, next) => {
+  const url = req.protocol + "://" + req.get("host");
+  const post = new Post({
+    title: req.body.title,
+    content: req.body.content,
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId,
+  });
+  post
+    .save()
+    .then((createdPost) => {
+      res.status(201).json({
+        message: "Post Added Successfully",
+        post: {
+          ...createdPost,
+          id: createdPost._id,
+        },
+      });
+    })
+    .catch((error) => {
+      res.status(500).json({ message: "Post Creation Failed" });
+    });
+};
+
+exports.updatePost = (req, res, next) => {
+  let imagePath = req.body.imagePath;
+  if (req.file) {
+    const url = req.protocol + "://" + req.get("host");
+    imagePath = url + "/images/" + req.file.filename;
+  }
+  const post = new Post({
+    _id: req.body.id,
+    title: req.body.title,
+    content: req.body.content,
+    imagePath: imagePath,
+    creator: req.userData.userId,
+  });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
+    .then((result) => {
+      if (result.nModified > 0) {
+        console.log(result);
+        res.status(200).json({ message: "Update Successful!" });
+      } else {
+        res
+          .status(401)
+          .json({ message: "Update Not Successful! as Not Authorized" });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ message: "Could not update Post" });
+    });
+};
+
+exports.deletePost = (req, res, next) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
+    .then((result) => {
+      if (result.n > 0) {
+        console.log(result);
+        res.status(200).json({ message: "Delete Successful!" });
+      } else {
+        res
+          .status(401)
+          .json({ message: "Delete Not Successful! as Not Authorized" });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ message: "Could not update Post" });
+    });
+
+  res.status(200).json();
+};
+
+exports.getPost = (req, res, next) => {
+  Post.findById(req.params.id)
+    .then((post) => {
+      if (post) {
+        res.status(200).json(post);
+      } else {
+        res.status(404).json({ message: "Post Not Found" });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ message: "Can not get Posts" });
+    });
+};
+
+exports.getPosts = (req, res, next) => {
+  //+ convert To Number
+  const pageSize = +req.query.pagesize;
+  const currentPage = +req.query.page;
+  const postQuery = Post.find();
+  let fetchedPosts;
+  if (pageSize && currentPage) {
+    postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
+  }
+  postQuery
+    .then((documents) => {
+      fetchedPosts = documents;
+      return Post.collection.countDocuments();
+    })
+    .then((count) => {
+      return res.status(200).json({
+        message: "Posts Fetched Successfully",
+        posts: fetchedPosts,
+        maxPosts: count,
+      });
+    })
+    .catch((error) => {
+      res.status(500).json({ message: "Can not get Posts" });
+    });
+};
